server/requirements.txt
flask==2.1.0
flask-cors==3.0.10
openai-whisper @ git+https://github.com/openai/whisper.git
torch==1.13.1
transformers==4.25.1
scipy==1.10.0